SERVER-72469 Add integration tests for compound wildcard indexes

+++ jstests/core/index/wildcard/wildcard_index_type.js
@@ -2,48 +2,68 @@
  * Test $** support for the $type operator.
  * @tags: [
  *   assumes_read_concern_local,
+ *   does_not_support_stepdowns,
  * ]
  */
 (function() {
 "use strict";
 
-load("jstests/libs/analyze_plan.js");  // For getPlanStages.
+load("jstests/libs/analyze_plan.js");       // For getPlanStages.
+load("jstests/libs/feature_flag_util.js");  // For "FeatureFlagUtil"
 
 const coll = db.wildcard_index_type;
 coll.drop();
 
-const indexWildcard = {
-    "$**": 1
-};
+const wildcardIndexes = [
+    {keyPattern: {"$**": 1}},
+    {keyPattern: {"$**": 1, "other": 1}, wildcardProjection: {"other": 0}}
+];
+
+// TODO SERVER-68303: Remove the feature flag and update corresponding tests.
+const allowCompoundWildcardIndexes =
+    FeatureFlagUtil.isPresentAndEnabled(db.getMongo(), "CompoundWildcardIndexes");
 
 // Inserts the given document and runs the given query to confirm that:
 // (1) query matches the given document if match is true,
 // (2) the winning plan does a wildcard index scan, and
 // (3) the resulting index bound matches 'expectedBounds' if given.
 function assertExpectedDocAnswersWildcardIndexQuery(doc, query, match, expectedBounds) {
-    coll.drop();
-    assert.commandWorked(coll.createIndex(indexWildcard));
-    assert.commandWorked(coll.insert(doc));
+    for (const indexSpec of wildcardIndexes) {
+        if (!allowCompoundWildcardIndexes && indexSpec.wildcardProjection) {
+            continue;
+        }
+        coll.drop();
+        const option = {};
+        if (indexSpec.wildcardProjection) {
+            option['wildcardProjection'] = indexSpec.wildcardProjection;
+        }
 
-    // Check that a wildcard index scan is being used to answer query.
-    const explain = coll.explain("executionStats").find(query).finish();
-    if (!match) {
-        assert.eq(0, explain.executionStats.nReturned, explain);
-        return;
-    }
+        assert.commandWorked(coll.createIndex(indexSpec.keyPattern, option));
+        assert.commandWorked(coll.insert(doc));
 
-    // Check that the query returns the document.
-    assert.eq(1, explain.executionStats.nReturned, explain);
+        // Check that a wildcard index scan is being used to answer query.
+        const explain = coll.explain("executionStats").find(query).finish();
+        if (!match) {
+            assert.eq(0, explain.executionStats.nReturned, explain);
+            return;
+        }
 
-    // Winning plan uses a wildcard index scan.
-    const winningPlan = getWinningPlan(explain.queryPlanner);
-    const ixScans = getPlanStages(winningPlan, "IXSCAN");
-    assert.gt(ixScans.length, 0, explain);
-    ixScans.forEach((ixScan) => assert(ixScan.keyPattern.$_path));
+        // Check that the query returns the document.
+        assert.eq(1, explain.executionStats.nReturned, explain);
 
-    // Expected bounds were used.
-    if (expectedBounds !== undefined) {
-        ixScans.forEach((ixScan) => assert.docEq(expectedBounds, ixScan.indexBounds));
+        // Winning plan uses a wildcard index scan.
+        const winningPlan = getWinningPlan(explain.queryPlanner);
+        const ixScans = getPlanStages(winningPlan, "IXSCAN");
+        assert.gt(ixScans.length, 0, explain);
+        ixScans.forEach((ixScan) => assert(ixScan.keyPattern.$_path));
+
+        // Expected bounds were used.
+        if (expectedBounds !== undefined) {
+            if (indexSpec.keyPattern.other) {
+                expectedBounds['other'] = [`[MinKey, MaxKey]`];
+            }
+            ixScans.forEach((ixScan) => assert.docEq(expectedBounds, ixScan.indexBounds));
+        }
     }
 }
