+++ jstests/core/wildcard_index_dup_predicates.js
@@ -7,38 +7,54 @@
 (function() {
 "use strict";
 
-load("jstests/libs/analyze_plan.js");  // For getPlanStages.
+load("jstests/libs/analyze_plan.js");       // For getPlanStages.
+load("jstests/libs/feature_flag_util.js");  // For "FeatureFlagUtil"
 
 const coll = db.wildcard_index_dup_predicates;
 coll.drop();
 
-const indexWildcard = {
-    "$**": 1
-};
+const wildcardIndexes = [
+    {keyPattern: {"$**": 1}},
+    {keyPattern: {"$**": 1, c: 1}, wildcardProjection: {c: 0}},
+    {keyPattern: {"$**": -1, c: 1}, wildcardProjection: {c: 0}}
+];
+
+// TODO SERVER-68303: Remove the feature flag and update corresponding tests.
+const allowCompoundWildcardIndexes =
+    FeatureFlagUtil.isPresentAndEnabled(db.getMongo(), "CompoundWildcardIndexes");
 
 // Inserts the given document and runs the given query to confirm that:
 // (1) query matches the given document
 // (2) the winning plan does a wildcard index scan
 function assertExpectedDocAnswersWildcardIndexQuery(doc, query, match) {
-    coll.drop();
-    assert.commandWorked(coll.createIndex(indexWildcard));
-    assert.commandWorked(coll.insert(doc));
+    for (const indexSpec of wildcardIndexes) {
+        if (!allowCompoundWildcardIndexes && indexSpec.wildcardProjection) {
+            continue;
+        }
+        coll.drop();
+        const option = {};
+        if (indexSpec.wildcardProjection) {
+            option['wildcardProjection'] = indexSpec.wildcardProjection;
+        }
+        assert.commandWorked(coll.createIndex(indexSpec.keyPattern, option));
+        assert.commandWorked(coll.insert(doc));
 
-    // Check that a wildcard index scan is being used to answer query.
-    const explain = coll.explain("executionStats").find(query).finish();
-    if (!match) {
-        assert.eq(0, explain.executionStats.nReturned, explain);
-        return;
+        // Check that a wildcard index scan is being used to answer query.
+        const explain = coll.explain("executionStats").find(query).finish();
+        if (!match) {
+            assert.eq(0, explain.executionStats.nReturned, explain);
+            return;
+        }
+
+        // Check that the query returns the document.
+        assert.eq(1, explain.executionStats.nReturned, explain);
+
+        // Winning plan uses a wildcard index scan.
+        const winningPlan = getWinningPlan(explain.queryPlanner);
+        const ixScans = getPlanStages(winningPlan, "IXSCAN");
+        assert.gt(ixScans.length, 0, explain);
+        ixScans.forEach((ixScan) => assert.eq(true, ixScan.indexName.includes("$**")));
     }
-
-    // Check that the query returns the document.
-    assert.eq(1, explain.executionStats.nReturned, explain);
-
-    // Winning plan uses a wildcard index scan.
-    const winningPlan = getWinningPlan(explain.queryPlanner);
-    const ixScans = getPlanStages(winningPlan, "IXSCAN");
-    assert.gt(ixScans.length, 0, explain);
-    ixScans.forEach((ixScan) => assert.eq(ixScan.indexName, "$**_1"));
 }
 
 assertExpectedDocAnswersWildcardIndexQuery(
